+++ es6module/jspanel.js
@@ -1,6 +1,6 @@
 /**
  * jsPanel - A JavaScript library to create highly configurable multifunctional floating panels that can also be used as modal, tooltip, hint or contextmenu
- * @version v4.14.0
+ * @version v4.14.1
  * @homepage https://jspanel.de/
  * @license MIT
  * @author Stefan Sträßer - info@jspanel.de
@@ -11,8 +11,8 @@ export // eslint-disable-next-line no-redeclare
 // noinspection JSVoidFunctionReturnValueUsed
 // eslint-disable-next-line no-redeclare
 let jsPanel = {
-    version: '4.14.0',
-    date: '2022-04-20 15:57',
+    version: '4.14.1',
+    date: '2022-05-17 15:53',
     ajaxAlwaysCallbacks: [],
     autopositionSpacing: 4,
     closeOnEscape: (() => {
@@ -1887,7 +1887,6 @@ let jsPanel = {
                 values.height += 'px';
             }
         }
-
         return values; // return value must be an object {width: xxx, height: xxx}
     },
     pOborder(border) {
@@ -2477,7 +2476,6 @@ let jsPanel = {
                             jsPanel.perceivedBrightness(themeDetails.filling) <= jsPanel.colorBrightnessThreshold ? '#fff' : '#000';
                 }
             }
-
             return self;
         };
         self.applyCustomTheme = (theme) => {
@@ -4008,6 +4006,18 @@ let jsPanel = {
                                     );
                                 }
                                 self.front();
+
+                                // if panel is maximized on resize start set status to normalized and swap maximize/normalize buttons
+                                if (self.status === 'maximized') {
+                                    self.status = 'normalized';
+                                    if (self.controlbar.querySelector('.jsPanel-btn-maximize')) {
+                                        self.setControlStatus('maximize', 'show');
+                                    }
+                                    if (self.controlbar.querySelector('.jsPanel-btn-normalize')) {
+                                        self.setControlStatus('normalize', 'hide');
+                                    }
+                                }
+
                                 if (elmtRect.height > startHeight + 5) {
                                     self.status = 'normalized';
                                     self.setControls(['.jsPanel-btn-normalize']);