+++ source/jspanel.js
@@ -2,8 +2,8 @@
 // noinspection JSVoidFunctionReturnValueUsed
 // eslint-disable-next-line no-redeclare
 let jsPanel = {
-    version: '4.14.0',
-    date: '2022-04-20 15:57',
+    version: '4.14.1',
+    date: '2022-05-17 15:53',
     ajaxAlwaysCallbacks: [],
     autopositionSpacing: 4,
     closeOnEscape: (() => {
@@ -1878,7 +1878,6 @@ let jsPanel = {
                 values.height += 'px';
             }
         }
-
         return values; // return value must be an object {width: xxx, height: xxx}
     },
     pOborder(border) {
@@ -2468,7 +2467,6 @@ let jsPanel = {
                             jsPanel.perceivedBrightness(themeDetails.filling) <= jsPanel.colorBrightnessThreshold ? '#fff' : '#000';
                 }
             }
-
             return self;
         };
         self.applyCustomTheme = (theme) => {
@@ -3999,6 +3997,18 @@ let jsPanel = {
                                     );
                                 }
                                 self.front();
+
+                                // if panel is maximized on resize start set status to normalized and swap maximize/normalize buttons
+                                if (self.status === 'maximized') {
+                                    self.status = 'normalized';
+                                    if (self.controlbar.querySelector('.jsPanel-btn-maximize')) {
+                                        self.setControlStatus('maximize', 'show');
+                                    }
+                                    if (self.controlbar.querySelector('.jsPanel-btn-normalize')) {
+                                        self.setControlStatus('normalize', 'hide');
+                                    }
+                                }
+
                                 if (elmtRect.height > startHeight + 5) {
                                     self.status = 'normalized';
                                     self.setControls(['.jsPanel-btn-normalize']);